backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
networkx==3.2.1
causal-learn==0.1.3.8
sdv==1.10.0
ctgan==0.9.0
gymnasium==0.29.1
stable-baselines3==2.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
duckdb==0.9.2
pyarrow==14.0.1
torch==2.1.1
matplotlib==3.8.2
seaborn==0.13.0
httpx==0.25.2
openpyxl==3.1.2
lingam